Data wrapper for analysed perturbation data using a hierarchical model

Class HMAnalysedPerturbationData is a wrapper for various objects of an analysis of a perturbation experiment done using a hierarchical model. Class HMAnalysedPerturbationData exposes getters for its members of the same name, but no setters, because the data should be treated as constant once set. Objects of class HMAnalysedPerturbationData do not need to be constructed manually but are returned from calling hm (see the examples).

Slots

geneHits

prioritized genes

nestedGeneEffects

the estimated effect sizes for genes on a viral level

nestedGeneHits

nested prioritized genes

modelFit

the fitted model

Examples

 data(rnaiscreen)
 res <- hm(rnaiscreen, effect.size=0.01)

#> Fitting hierarchical model

 class(res)

#> [1] "HMAnalysedPerturbationData"
#> attr(,"package")
#> [1] "perturbatr"