Data wrapper for analysed perturbation data using network diffusion

Class NetworkAnalysedPerturbationData is a wrapper for various objects of an analysis of a perturbation experiment done using network diffusion. Class NetworkAnalysedPerturbationData exposes getters for its members of the same name, but no setters, because the data should be treated as constant once set. Objects of class NetworkAnalysedPerturbationData do not need to be constructed manually but are returned from calling diffuse (see the examples).

Slots

initialModel

the model that was provided for analysis

graph

an igraph object that served for the diffusion process

Examples

 data(rnaiscreen)
 hm.fit <- hm(rnaiscreen)

#> Fitting hierarchical model

 graph <- readRDS(system.file(
   "extdata", "graph_small.rds", package = "perturbatr"))
 res <- diffuse(hm.fit, graph=graph, r=1)

#> Only taking largest connected component to ensure ergodicity.

#> Diffusion using Markov random walks.

#> Correcting for hub degrees.

#> normalizing column vectors!

#> normalizing column vectors!