Data wrapper for a data set of a perturbation screen.

Class PerturbationData wraps a data set derived from a genetic perturbation screen, e.g. using RNA interference or CRISPR. Class PerturbationData exposes getters for its members of the same name, but no setters, because the data should be treated as constant once set. The easiest way to construct a PerturbationData object is by first creating a data.frame and then calling as. See the examples to construct an object.

Slots

dataSet

the data set as a tibble

Examples

  df <- data.frame(Condition    = c("V1", "V2", "V3"),
                   Replicate    = c(1, 1, 1),
                   GeneSymbol   = c("TP52", "NegCtrl", "PosCtrl"),
                   Perturbation = c("P1", "P2", "P3"),
                   Readout      = c(123, 121, 12),
                   Control      = c(0, -1, 1))
  methods::as(df, "PerturbationData")

#> A perturbation data set
#> 
#>   Condition GeneSymbol Readout
#> 1        V1       TP52     123
#> 2        V1       TP52     123
#> 3        V2    NegCtrl     121
#> 4        V2    NegCtrl     121
#> 5        V3    PosCtrl      12
#> 6        V3    PosCtrl      12