Network diffusion

Propagate the estimated gene effects from a previous analysis over a network using network diffusion. First the estimated effects are normalized and mapped to a given genetic network, for instance a PPI or co-expression network. Then the normalized effects are propagated across the edges of the network using a Markov random walk with restarts. By that the initial ranking of genes (as given by their absolute effect sizes) is re-evaluated and the genes are reordered. Thus network diffusion potentially reduced false negative hits.

diffuse(obj, path = NULL, graph = NULL, r = 0.5,
  delete.nodes.on.degree = 0, do.bootstrap = FALSE)

# S4 method for HMAnalysedPerturbationData
diffuse(obj, path = NULL,
  graph = NULL, r = 0.5, delete.nodes.on.degree = 0,
  do.bootstrap = FALSE)

Arguments

obj
HMAnalysedPerturbationData object

path
path to the network file (if graph is NULL)

graph
an weighted adjacency matrix (if path is NULL)

r
restart probability of the random walk

delete.nodes.on.degree
delete nodes from the graph with a degree of less or equal than delete.nodes.on.degree

do.bootstrap
run a diffusion on every bootstrap sample in case bootstrap samples are available

Value

returns a NetworkAnalysedPerturbationData object

Examples

 data(rnaiscreen)
 res        <- hm(rnaiscreen, effect.size=0.01)

#> Fitting hierarchical model

 graph.file <- system.file("extdata", "graph_file.tsv",
                            package = "perturbatr")
 diffu      <- diffuse(res, path=graph.file, r=1)

#> Only taking largest connected component to ensure ergodicity.

#> Diffusion using Markov random walks.

#> normalizing column vectors!