Filter the rows of a perturbation data set

Takes a perturbation data set and filters the rows by some criterion. The filtered object will have the same type as the previous object.

filter(obj, ...)

Arguments

obj
the object to be filtered

...
variable number of logical predicates in terms of the column names in obj. Multiple predicates will be combined with a logical 'and'. Rows where all conditions are met will be kept. The column names do not need to be quoted. Wraps around dplyr::filter.

Value

returns an object of the same type filtered by some criteria

Examples

 data(rnaiscreen)
 flt.dat <- filter(rnaiscreen, Condition=="V1")
 flt.dat <- filter(rnaiscreen, Condition=="V1", RowIdx==1)