Getter for graph used for network diffusion

Returns the graph that has been used for analysis using network diffusion.

graph(obj)

# S4 method for NetworkAnalysedPerturbationData
graph(obj)

Arguments

obj
the object for which you want to extract the underlying graph

Value

returns a igraph object

Examples

 data(rnaiscreen)
 graph.file <- system.file("extdata", "graph_file.tsv",
                            package = "perturbatr")
 ft <- hm(rnaiscreen, effect.size=0.01)

#> Fitting hierarchical model

 diffu <- diffuse(ft, path=graph.file, r=1)

#> Only taking largest connected component to ensure ergodicity.

#> Diffusion using Markov random walks.

#> normalizing column vectors!

 graph(diffu)

#> IGRAPH 2a0e6e9 UNW- 128 239 -- 
#> + attr: name (v/c), weight (e/n)
#> + edges from 2a0e6e9 (vertex names):
#>  [1] cacna1s--cav3           cd247  --cd3e           cav3   --cd63          
#>  [4] cd247  --cd8b           cd3e   --cd8b           aurka  --cep131        
#>  [7] aurka  --chek1          aurka  --chuk           ank3   --cog6          
#> [10] ank3   --cog8           cog6   --cog8           ank3   --copz1         
#> [13] cog6   --copz1          cog8   --copz1          apc    --cul1          
#> [16] chek1  --cul1           chuk   --cul1           ank3   --dctn2         
#> [19] aurka  --dctn2          cep131 --dctn2          cog6   --dctn2         
#> [22] cog8   --dctn2          copz1  --dctn2          chuk   --dkfzp686j04131
#> + ... omitted several edges