Getter for boolean if boostrapping was used

Returns a boolean if for the analysed object bootstrapping was used to create confidence intervals.

isBootstrapped(obj)

# S4 method for AbstractAnalysedPerturbationData
isBootstrapped(obj)

Arguments

obj
the object for which you want to extract the boolean

Value

returns a boolean.

Examples

 data(rnaiscreen)
 ft <- hm(rnaiscreen)

#> Fitting hierarchical model

 isBootstrapped(ft)

#> [1] FALSE