Getter for model fit

Returns the mixed effects model fit from the analysis using a hierarchical model.

modelFit(obj)

# S4 method for HMAnalysedPerturbationData
modelFit(obj)

Arguments

obj
the object for which you want to extract the underlying fit

Value

returns a list

Examples

 data(rnaiscreen)
 ft <- hm(rnaiscreen)

#> Fitting hierarchical model

 modelFit(ft)

#> $fit
#> Linear mixed model fit by REML ['lmerMod']
#> Formula: Readout ~ Condition + (1 | GeneSymbol) + (1 | Condition:GeneSymbol)
#>    Data: md
#> Weights: md$Weight
#> REML criterion at convergence: 379378.4
#> Random effects:
#>  Groups               Name        Std.Dev.
#>  Condition:GeneSymbol (Intercept)   0.0   
#>  GeneSymbol           (Intercept) 135.8   
#>  Residual                         987.1   
#> Number of obs: 22800, groups:  Condition:GeneSymbol, 760; GeneSymbol, 380
#> Fixed Effects:
#> (Intercept)  ConditionV2  
#>   9.968e+02    6.536e-12  
#> 
#> $ge.fdrs
#> $ge.fdrs$ret
#> # A tibble: 380 x 2
#>    GeneSymbol  Qval
#>    <chr>      <dbl>
#>  1 abcb10        NA
#>  2 acadl         NA
#>  3 adam21        NA
#>  4 adamts7       NA
#>  5 adh6          NA
#>  6 adra2c        NA
#>  7 alk           NA
#>  8 ank3          NA
#>  9 ano5          NA
#> 10 anos1         NA
#> # ... with 370 more rows
#> 
#> $ge.fdrs$btst
#> [1] FALSE
#> 
#> 
#> $nge.fdrs
#> $nge.fdrs$nges
#>         GeneSymbol           V1           V2 Qval.V1 Qval.V2
#> 1           abcb10   85.8439806   85.8439806      NA      NA
#> 2            acadl    1.0550429    1.0550429      NA      NA
#> 3           adam21  109.2178510  109.2178510      NA      NA
#> 4          adamts7   74.1409826   74.1409826      NA      NA
#> 5             adh6  -39.7663590  -39.7663590      NA      NA
#> 6           adra2c  143.3751409  143.3751409      NA      NA
#> 7              alk  -34.6897248  -34.6897248      NA      NA
#> 8             ank3   86.9971951   86.9971951      NA      NA
#> 9             ano5  -35.3023219  -35.3023219      NA      NA
#> 10           anos1   85.6947369   85.6947369      NA      NA
#> 11             apc    8.8051713    8.8051713      NA      NA
#> 12            araf -135.9541491 -135.9541491      NA      NA
#> 13        arhgap30   -1.6088912   -1.6088912      NA      NA
#> 14        arhgap33  -18.4547788  -18.4547788      NA      NA
#> 15        arhgap44  189.6569506  189.6569506      NA      NA
#> 16        arhgef28   32.2642041   32.2642041      NA      NA
#> 17          arid1b  -46.1098118  -46.1098118      NA      NA
#> 18          arpc5l   77.7889828   77.7889828      NA      NA
#> 19            arsk  150.8823131  150.8823131      NA      NA
#> 20           asah2  -24.3459009  -24.3459009      NA      NA
#> 21           asb12 -100.3695360 -100.3695360      NA      NA
#> 22           atcay   66.8565721   66.8565721      NA      NA
#> 23          atp1b1  -47.9105370  -47.9105370      NA      NA
#> 24          atp2a3  -56.1632787  -56.1632787      NA      NA
#> 25          atp5j2   -7.4948990   -7.4948990      NA      NA
#> 26          atpif1 -153.9780391 -153.9780391      NA      NA
#> 27         atxn7l3   13.0379768   13.0379768      NA      NA
#> 28           aurka  -93.9806250  -93.9806250      NA      NA
#> 29          b3gnt6   13.9773285   13.9773285      NA      NA
#> 30          b4gat1   41.5963000   41.5963000      NA      NA
#> 31           baz2a  126.6317311  126.6317311      NA      NA
#> 32            bcam   96.1319445   96.1319445      NA      NA
#> 33           bccip  159.0925063  159.0925063      NA      NA
#> 34           brwd3   23.4660534   23.4660534      NA      NA
#> 35            btla -189.9286097 -189.9286097      NA      NA
#> 36            c1qb  -79.4799530  -79.4799530      NA      NA
#> 37              c3  -44.6337482  -44.6337482      NA      NA
#> 38         cacna1e -168.5766267 -168.5766267      NA      NA
#> 39         cacna1s   61.5219528   61.5219528      NA      NA
#> 40           cadm1  -59.4840814  -59.4840814      NA      NA
#> 41           calcr -119.1386853 -119.1386853      NA      NA
#> 42           camlg -119.3392683 -119.3392683      NA      NA
#> 43          card10  -43.8233469  -43.8233469      NA      NA
#> 44          card17  -69.0268204  -69.0268204      NA      NA
#> 45            cav3   66.9058348   66.9058348      NA      NA
#> 46          cct8l2 -126.1172608 -126.1172608      NA      NA
#> 47          cd300e   31.5709743   31.5709743      NA      NA
#> 48            cd3e  -24.9584774  -24.9584774      NA      NA
#> 49            cd63  -87.7905470  -87.7905470      NA      NA
#> 50            cd8b   66.6051796   66.6051796      NA      NA
#> 51           cdca5  204.3381073  204.3381073      NA      NA
#> 52           cdk10   75.9841078   75.9841078      NA      NA
#> 53            cdk6 -103.7357829 -103.7357829      NA      NA
#> 54           cebpe  186.2732155  186.2732155      NA      NA
#> 55          cep131    8.5113717    8.5113717      NA      NA
#> 56            chd1   13.0531852   13.0531852      NA      NA
#> 57           chd1l -149.9958471 -149.9958471      NA      NA
#> 58           chek1 -194.5641699 -194.5641699      NA      NA
#> 59           chmp7  167.5428226  167.5428226      NA      NA
#> 60           chpf2  151.3503424  151.3503424      NA      NA
#> 61           chrm3  127.6826656  127.6826656      NA      NA
#> 62           chst8 -111.4987498 -111.4987498      NA      NA
#> 63            chuk  -65.8075837  -65.8075837      NA      NA
#> 64            clk2 -117.9729693 -117.9729693      NA      NA
#> 65            clpx  -60.0093097  -60.0093097      NA      NA
#> 66          clstn1  -24.4240077  -24.4240077      NA      NA
#> 67            cog6   53.9418471   53.9418471      NA      NA
#> 68            cog8  -67.1117409  -67.1117409      NA      NA
#> 69           cops8 -225.0675015 -225.0675015      NA      NA
#> 70           copz1 -124.5989255 -124.5989255      NA      NA
#> 71           cotl1    1.7374631    1.7374631      NA      NA
#> 72            cptp -179.4751154 -179.4751154      NA      NA
#> 73           cradd -210.4428095 -210.4428095      NA      NA
#> 74         creb3l2  133.1496892  133.1496892      NA      NA
#> 75           crhbp   18.4808508   18.4808508      NA      NA
#> 76             crx  215.5555143  215.5555143      NA      NA
#> 77          csf2rb  -36.7822229  -36.7822229      NA      NA
#> 78          ctnnd1  -80.3673519  -80.3673519      NA      NA
#> 79            cul1  -14.6635770  -14.6635770      NA      NA
#> 80          d2hgdh  107.7899207  107.7899207      NA      NA
#> 81            dad1  -54.8530415  -54.8530415      NA      NA
#> 82           dctn2  -14.2478440  -14.2478440      NA      NA
#> 83         dcun1d5 -153.9181554 -153.9181554      NA      NA
#> 84           ddx49   51.8196395   51.8196395      NA      NA
#> 85            dgkq  -19.9743799  -19.9743799      NA      NA
#> 86           dhx58  156.2689426  156.2689426      NA      NA
#> 87          diablo -124.8705544 -124.8705544      NA      NA
#> 88  dkfzp686j04131  124.7027978  124.7027978      NA      NA
#> 89  dkfzp686k04147  -59.6191987  -59.6191987      NA      NA
#> 90   dkfzp686o0215    2.9709083    2.9709083      NA      NA
#> 91    dkfzp781d102    5.3138304    5.3138304      NA      NA
#> 92            dlx4  -62.6180304  -62.6180304      NA      NA
#> 93         dnajc5b  -18.6012915  -18.6012915      NA      NA
#> 94           dock6  -55.7522391  -55.7522391      NA      NA
#> 95             dst  -61.7678879  -61.7678879      NA      NA
#> 96          dyrk1b  -62.2944422  -62.2944422      NA      NA
#> 97            e2f8  156.7219419  156.7219419      NA      NA
#> 98            edc3   38.9247574   38.9247574      NA      NA
#> 99           edrf1  271.7072242  271.7072242      NA      NA
#> 100            eed  -27.3160226  -27.3160226      NA      NA
#> 101         eef1e1   40.5082890   40.5082890      NA      NA
#> 102          efna1  -84.8993644  -84.8993644      NA      NA
#> 103         eftud2    4.6106512    4.6106512      NA      NA
#> 104        eif2ak4  -52.3864507  -52.3864507      NA      NA
#> 105         eif2b3  -11.9159429  -11.9159429      NA      NA
#> 106          eif3g  -18.1246695  -18.1246695      NA      NA
#> 107         eif4g2 -141.7724843 -141.7724843      NA      NA
#> 108           elf1  -68.3328215  -68.3328215      NA      NA
#> 109         elovl1 -169.5953783 -169.5953783      NA      NA
#> 110         elovl3  124.9715179  124.9715179      NA      NA
#> 111         elovl4 -122.3005131 -122.3005131      NA      NA
#> 112           emsy  165.3245041  165.3245041      NA      NA
#> 113        epb41l1  -30.7018806  -30.7018806      NA      NA
#> 114           eprs  -45.5879984  -45.5879984      NA      NA
#> 115          ethe1   61.2073643   61.2073643      NA      NA
#> 116           etv2  -75.8546317  -75.8546317      NA      NA
#> 117           evc2  112.9134878  112.9134878      NA      NA
#> 118         exosc7    2.9527499    2.9527499      NA      NA
#> 119           f13b  142.7175588  142.7175588      NA      NA
#> 120         faap20   20.1336756   20.1336756      NA      NA
#> 121           fat4 -107.7585698 -107.7585698      NA      NA
#> 122           fbf1  -12.9595575  -12.9595575      NA      NA
#> 123         fbxo18  -19.6884189  -19.6884189      NA      NA
#> 124           fcar   44.8455933   44.8455933      NA      NA
#> 125         fip1l1   19.2967520   19.2967520      NA      NA
#> 126       flj00153   11.1017110   11.1017110      NA      NA
#> 127          flrt3   27.5069616   27.5069616      NA      NA
#> 128         flvcr1  -34.8533945  -34.8533945      NA      NA
#> 129          frat1   31.5561556   31.5561556      NA      NA
#> 130            fus   76.5190648   76.5190648      NA      NA
#> 131           gab1  -47.0928761  -47.0928761      NA      NA
#> 132          gata1  -96.6371903  -96.6371903      NA      NA
#> 133           gbf1 -109.5108795 -109.5108795      NA      NA
#> 134           gbp1   98.1838105   98.1838105      NA      NA
#> 135         gdpgp1  196.4408059  196.4408059      NA      NA
#> 136            gip  -11.4165314  -11.4165314      NA      NA
#> 137           gli2   75.0993307   75.0993307      NA      NA
#> 138           glrb  -21.6673020  -21.6673020      NA      NA
#> 139          glrx3   46.5600541   46.5600541      NA      NA
#> 140          grik4 -166.7899939 -166.7899939      NA      NA
#> 141         h2afb2   18.4019227   18.4019227      NA      NA
#> 142           hagh  124.0505439  124.0505439      NA      NA
#> 143           hat1  -79.0342200  -79.0342200      NA      NA
#> 144      hel-s-109 -107.8964118 -107.8964118      NA      NA
#> 145        hel-s-4 -166.4665086 -166.4665086      NA      NA
#> 146           hes5   86.3759043   86.3759043      NA      NA
#> 147           hic2    9.2051715    9.2051715      NA      NA
#> 148      hist1h2bn -177.7925224 -177.7925224      NA      NA
#> 149        hist3h3   31.5748053   31.5748053      NA      NA
#> 150         hivep2   53.9894112   53.9894112      NA      NA
#> 151          hla-b   92.9372726   92.9372726      NA      NA
#> 152        hla-dob  -41.3309576  -41.3309576      NA      NA
#> 153          hmgn3  -30.3565748  -30.3565748      NA      NA
#> 154          hnf4g  -52.9694503  -52.9694503      NA      NA
#> 155        hnrnpa1  249.4615566  249.4615566      NA      NA
#> 156         hoxa13  -58.2388060  -58.2388060      NA      NA
#> 157         hoxd12  -18.8031735  -18.8031735      NA      NA
#> 158          hsfx1 -134.7892470 -134.7892470      NA      NA
#> 159          hsph1  -81.2591325  -81.2591325      NA      NA
#> 160           htn1 -238.1124628 -238.1124628      NA      NA
#> 161          htr3e   33.2699653   33.2699653      NA      NA
#> 162            id3  100.1220673  100.1220673      NA      NA
#> 163          ifna7  -43.4091640  -43.4091640      NA      NA
#> 164          ifnb1   32.4899353   32.4899353      NA      NA
#> 165         ift140  136.2961306  136.2961306      NA      NA
#> 166           il19   86.7683414   86.7683414      NA      NA
#> 167         ino80c  -55.1174980  -55.1174980      NA      NA
#> 168          insm1  115.4525754  115.4525754      NA      NA
#> 169           insr   10.3341920   10.3341920      NA      NA
#> 170          ints7 -132.1912109 -132.1912109      NA      NA
#> 171           intu   37.0731078   37.0731078      NA      NA
#> 172           ipo7  160.5005424  160.5005424      NA      NA
#> 173          irs-1 -108.3122067 -108.3122067      NA      NA
#> 174           isy1  -47.1672629  -47.1672629      NA      NA
#> 175           kat5   30.0122107   30.0122107      NA      NA
#> 176          kcne1   63.0449310   63.0449310      NA      NA
#> 177          kcnh5    3.3894391    3.3894391      NA      NA
#> 178          kcnh8  -14.5194437  -14.5194437      NA      NA
#> 179          kcnq2  -94.4093575  -94.4093575      NA      NA
#> 180           kera   33.7502208   33.7502208      NA      NA
#> 181       kiaa1161  147.7109482  147.7109482      NA      NA
#> 182        kir2dl3 -113.0670367 -113.0670367      NA      NA
#> 183       kir2dl5a  -44.5773899  -44.5773899      NA      NA
#> 184             kl  100.5481162  100.5481162      NA      NA
#> 185            klb   15.8401055   15.8401055      NA      NA
#> 186          klhl9   38.0492411   38.0492411      NA      NA
#> 187           klk6    5.0147928    5.0147928      NA      NA
#> 188          krt19  -69.2039579  -69.2039579      NA      NA
#> 189           lcat    2.9386985    2.9386985      NA      NA
#> 190           lcp2 -154.0657226 -154.0657226      NA      NA
#> 191           ldb3   73.9494104   73.9494104      NA      NA
#> 192           lgr6 -150.2455229 -150.2455229      NA      NA
#> 193           lhx1  -42.3092126  -42.3092126      NA      NA
#> 194          lims1  100.3087639  100.3087639      NA      NA
#> 195          lpar4   42.2196445   42.2196445      NA      NA
#> 196          lrch4  138.4317185  138.4317185      NA      NA
#> 197           lyl1  150.4857196  150.4857196      NA      NA
#> 198         map2k1  -68.8996995  -68.8996995      NA      NA
#> 199        map3k14  137.6424289  137.6424289      NA      NA
#> 200         map3k3 -114.7746412 -114.7746412      NA      NA
#>  [ reached getOption("max.print") -- omitted 180 rows ]
#> 
#> $nge.fdrs$fdrs
#> list()
#> 
#> $nge.fdrs$nested.gene.matrix
#>         GeneSymbol Condition       Effect Qval
#> 1           abcb10        V1   85.8439806   NA
#> 2            acadl        V1    1.0550429   NA
#> 3           adam21        V1  109.2178510   NA
#> 4          adamts7        V1   74.1409826   NA
#> 5             adh6        V1  -39.7663590   NA
#> 6           adra2c        V1  143.3751409   NA
#> 7              alk        V1  -34.6897248   NA
#> 8             ank3        V1   86.9971951   NA
#> 9             ano5        V1  -35.3023219   NA
#> 10           anos1        V1   85.6947369   NA
#> 11             apc        V1    8.8051713   NA
#> 12            araf        V1 -135.9541491   NA
#> 13        arhgap30        V1   -1.6088912   NA
#> 14        arhgap33        V1  -18.4547788   NA
#> 15        arhgap44        V1  189.6569506   NA
#> 16        arhgef28        V1   32.2642041   NA
#> 17          arid1b        V1  -46.1098118   NA
#> 18          arpc5l        V1   77.7889828   NA
#> 19            arsk        V1  150.8823131   NA
#> 20           asah2        V1  -24.3459009   NA
#> 21           asb12        V1 -100.3695360   NA
#> 22           atcay        V1   66.8565721   NA
#> 23          atp1b1        V1  -47.9105370   NA
#> 24          atp2a3        V1  -56.1632787   NA
#> 25          atp5j2        V1   -7.4948990   NA
#> 26          atpif1        V1 -153.9780391   NA
#> 27         atxn7l3        V1   13.0379768   NA
#> 28           aurka        V1  -93.9806250   NA
#> 29          b3gnt6        V1   13.9773285   NA
#> 30          b4gat1        V1   41.5963000   NA
#> 31           baz2a        V1  126.6317311   NA
#> 32            bcam        V1   96.1319445   NA
#> 33           bccip        V1  159.0925063   NA
#> 34           brwd3        V1   23.4660534   NA
#> 35            btla        V1 -189.9286097   NA
#> 36            c1qb        V1  -79.4799530   NA
#> 37              c3        V1  -44.6337482   NA
#> 38         cacna1e        V1 -168.5766267   NA
#> 39         cacna1s        V1   61.5219528   NA
#> 40           cadm1        V1  -59.4840814   NA
#> 41           calcr        V1 -119.1386853   NA
#> 42           camlg        V1 -119.3392683   NA
#> 43          card10        V1  -43.8233469   NA
#> 44          card17        V1  -69.0268204   NA
#> 45            cav3        V1   66.9058348   NA
#> 46          cct8l2        V1 -126.1172608   NA
#> 47          cd300e        V1   31.5709743   NA
#> 48            cd3e        V1  -24.9584774   NA
#> 49            cd63        V1  -87.7905470   NA
#> 50            cd8b        V1   66.6051796   NA
#> 51           cdca5        V1  204.3381073   NA
#> 52           cdk10        V1   75.9841078   NA
#> 53            cdk6        V1 -103.7357829   NA
#> 54           cebpe        V1  186.2732155   NA
#> 55          cep131        V1    8.5113717   NA
#> 56            chd1        V1   13.0531852   NA
#> 57           chd1l        V1 -149.9958471   NA
#> 58           chek1        V1 -194.5641699   NA
#> 59           chmp7        V1  167.5428226   NA
#> 60           chpf2        V1  151.3503424   NA
#> 61           chrm3        V1  127.6826656   NA
#> 62           chst8        V1 -111.4987498   NA
#> 63            chuk        V1  -65.8075837   NA
#> 64            clk2        V1 -117.9729693   NA
#> 65            clpx        V1  -60.0093097   NA
#> 66          clstn1        V1  -24.4240077   NA
#> 67            cog6        V1   53.9418471   NA
#> 68            cog8        V1  -67.1117409   NA
#> 69           cops8        V1 -225.0675015   NA
#> 70           copz1        V1 -124.5989255   NA
#> 71           cotl1        V1    1.7374631   NA
#> 72            cptp        V1 -179.4751154   NA
#> 73           cradd        V1 -210.4428095   NA
#> 74         creb3l2        V1  133.1496892   NA
#> 75           crhbp        V1   18.4808508   NA
#> 76             crx        V1  215.5555143   NA
#> 77          csf2rb        V1  -36.7822229   NA
#> 78          ctnnd1        V1  -80.3673519   NA
#> 79            cul1        V1  -14.6635770   NA
#> 80          d2hgdh        V1  107.7899207   NA
#> 81            dad1        V1  -54.8530415   NA
#> 82           dctn2        V1  -14.2478440   NA
#> 83         dcun1d5        V1 -153.9181554   NA
#> 84           ddx49        V1   51.8196395   NA
#> 85            dgkq        V1  -19.9743799   NA
#> 86           dhx58        V1  156.2689426   NA
#> 87          diablo        V1 -124.8705544   NA
#> 88  dkfzp686j04131        V1  124.7027978   NA
#> 89  dkfzp686k04147        V1  -59.6191987   NA
#> 90   dkfzp686o0215        V1    2.9709083   NA
#> 91    dkfzp781d102        V1    5.3138304   NA
#> 92            dlx4        V1  -62.6180304   NA
#> 93         dnajc5b        V1  -18.6012915   NA
#> 94           dock6        V1  -55.7522391   NA
#> 95             dst        V1  -61.7678879   NA
#> 96          dyrk1b        V1  -62.2944422   NA
#> 97            e2f8        V1  156.7219419   NA
#> 98            edc3        V1   38.9247574   NA
#> 99           edrf1        V1  271.7072242   NA
#> 100            eed        V1  -27.3160226   NA
#> 101         eef1e1        V1   40.5082890   NA
#> 102          efna1        V1  -84.8993644   NA
#> 103         eftud2        V1    4.6106512   NA
#> 104        eif2ak4        V1  -52.3864507   NA
#> 105         eif2b3        V1  -11.9159429   NA
#> 106          eif3g        V1  -18.1246695   NA
#> 107         eif4g2        V1 -141.7724843   NA
#> 108           elf1        V1  -68.3328215   NA
#> 109         elovl1        V1 -169.5953783   NA
#> 110         elovl3        V1  124.9715179   NA
#> 111         elovl4        V1 -122.3005131   NA
#> 112           emsy        V1  165.3245041   NA
#> 113        epb41l1        V1  -30.7018806   NA
#> 114           eprs        V1  -45.5879984   NA
#> 115          ethe1        V1   61.2073643   NA
#> 116           etv2        V1  -75.8546317   NA
#> 117           evc2        V1  112.9134878   NA
#> 118         exosc7        V1    2.9527499   NA
#> 119           f13b        V1  142.7175588   NA
#> 120         faap20        V1   20.1336756   NA
#> 121           fat4        V1 -107.7585698   NA
#> 122           fbf1        V1  -12.9595575   NA
#> 123         fbxo18        V1  -19.6884189   NA
#> 124           fcar        V1   44.8455933   NA
#> 125         fip1l1        V1   19.2967520   NA
#> 126       flj00153        V1   11.1017110   NA
#> 127          flrt3        V1   27.5069616   NA
#> 128         flvcr1        V1  -34.8533945   NA
#> 129          frat1        V1   31.5561556   NA
#> 130            fus        V1   76.5190648   NA
#> 131           gab1        V1  -47.0928761   NA
#> 132          gata1        V1  -96.6371903   NA
#> 133           gbf1        V1 -109.5108795   NA
#> 134           gbp1        V1   98.1838105   NA
#> 135         gdpgp1        V1  196.4408059   NA
#> 136            gip        V1  -11.4165314   NA
#> 137           gli2        V1   75.0993307   NA
#> 138           glrb        V1  -21.6673020   NA
#> 139          glrx3        V1   46.5600541   NA
#> 140          grik4        V1 -166.7899939   NA
#> 141         h2afb2        V1   18.4019227   NA
#> 142           hagh        V1  124.0505439   NA
#> 143           hat1        V1  -79.0342200   NA
#> 144      hel-s-109        V1 -107.8964118   NA
#> 145        hel-s-4        V1 -166.4665086   NA
#> 146           hes5        V1   86.3759043   NA
#> 147           hic2        V1    9.2051715   NA
#> 148      hist1h2bn        V1 -177.7925224   NA
#> 149        hist3h3        V1   31.5748053   NA
#> 150         hivep2        V1   53.9894112   NA
#> 151          hla-b        V1   92.9372726   NA
#> 152        hla-dob        V1  -41.3309576   NA
#> 153          hmgn3        V1  -30.3565748   NA
#> 154          hnf4g        V1  -52.9694503   NA
#> 155        hnrnpa1        V1  249.4615566   NA
#> 156         hoxa13        V1  -58.2388060   NA
#> 157         hoxd12        V1  -18.8031735   NA
#> 158          hsfx1        V1 -134.7892470   NA
#> 159          hsph1        V1  -81.2591325   NA
#> 160           htn1        V1 -238.1124628   NA
#> 161          htr3e        V1   33.2699653   NA
#> 162            id3        V1  100.1220673   NA
#> 163          ifna7        V1  -43.4091640   NA
#> 164          ifnb1        V1   32.4899353   NA
#> 165         ift140        V1  136.2961306   NA
#> 166           il19        V1   86.7683414   NA
#> 167         ino80c        V1  -55.1174980   NA
#> 168          insm1        V1  115.4525754   NA
#> 169           insr        V1   10.3341920   NA
#> 170          ints7        V1 -132.1912109   NA
#> 171           intu        V1   37.0731078   NA
#> 172           ipo7        V1  160.5005424   NA
#> 173          irs-1        V1 -108.3122067   NA
#> 174           isy1        V1  -47.1672629   NA
#> 175           kat5        V1   30.0122107   NA
#> 176          kcne1        V1   63.0449310   NA
#> 177          kcnh5        V1    3.3894391   NA
#> 178          kcnh8        V1  -14.5194437   NA
#> 179          kcnq2        V1  -94.4093575   NA
#> 180           kera        V1   33.7502208   NA
#> 181       kiaa1161        V1  147.7109482   NA
#> 182        kir2dl3        V1 -113.0670367   NA
#> 183       kir2dl5a        V1  -44.5773899   NA
#> 184             kl        V1  100.5481162   NA
#> 185            klb        V1   15.8401055   NA
#> 186          klhl9        V1   38.0492411   NA
#> 187           klk6        V1    5.0147928   NA
#> 188          krt19        V1  -69.2039579   NA
#> 189           lcat        V1    2.9386985   NA
#> 190           lcp2        V1 -154.0657226   NA
#> 191           ldb3        V1   73.9494104   NA
#> 192           lgr6        V1 -150.2455229   NA
#> 193           lhx1        V1  -42.3092126   NA
#> 194          lims1        V1  100.3087639   NA
#> 195          lpar4        V1   42.2196445   NA
#> 196          lrch4        V1  138.4317185   NA
#> 197           lyl1        V1  150.4857196   NA
#> 198         map2k1        V1  -68.8996995   NA
#> 199        map3k14        V1  137.6424289   NA
#> 200         map3k3        V1 -114.7746412   NA
#> 201        micall1        V1  -61.5276052   NA
#> 202            mlx        V1   78.9874868   NA
#> 203          mmrn1        V1 -100.5447374   NA
#> 204         mrpl15        V1   60.3554912   NA
#> 205          mrpl9        V1  -14.1297541   NA
#> 206          mrps5        V1 -209.3231847   NA
#> 207           msl1        V1   -5.0178542   NA
#> 208            msn        V1   55.8639254   NA
#> 209         mt-nd3        V1  -44.4129934   NA
#> 210          mtmr6        V1   53.5592816   NA
#> 211           mtpn        V1   88.1341218   NA
#> 212           mtrr        V1   12.6442689   NA
#> 213          myd88        V1 -159.6139568   NA
#> 214          myt1l        V1  110.4096857   NA
#> 215          naa20        V1   40.2856938   NA
#> 216          narfl        V1  -45.9447447   NA
#> 217           ncan        V1  -43.2428170   NA
#> 218           nck1        V1  -12.2651515   NA
#> 219           ncs1        V1   -2.3723015   NA
#> 220        ndufaf7        V1  -51.5772459   NA
#> 221          nedd1        V1   31.1933899   NA
#> 222         nelfcd        V1    4.9118303   NA
#> 223         nkx2-4        V1   19.7463728   NA
#> 224         nkx6-1        V1 -100.5911498   NA
#> 225         nmnat2        V1  -70.8149368   NA
#> 226           no66        V1   -1.5760866   NA
#> 227          npas4        V1  -99.0275193   NA
#> 228          nr3c1        V1   99.1300569   NA
#> 229            nsf        V1  -73.8204646   NA
#> 230          nsrp1        V1    6.5983720   NA
#> 231          nuak1        V1  140.5192113   NA
#> 232          nup93        V1   78.8564221   NA
#> 233         or10a2        V1  -64.7936607   NA
#> 234         or10h5        V1 -132.6513423   NA
#> 235         or13j1        V1 -171.7515740   NA
#> 236        or14c36        V1  -83.1862202   NA
#> 237          or1c1        V1  -82.0840808   NA
#> 238         or2b11        V1  136.4765741   NA
#> 239          or2m2        V1  225.3549227   NA
#> 240         or52a5        V1   17.3798751   NA
#> 241          or5f1        V1   83.7778881   NA
#> 242         or5m10        V1 -177.3439259   NA
#> 243        pcdhga2        V1   40.7996675   NA
#> 244           pck2        V1   -0.1946100   NA
#> 245          pde6g        V1   21.2956677   NA
#> 246          pdia2        V1  222.8484122   NA
#> 247           penk        V1  -30.8785953   NA
#> 248           per1        V1   95.3769586   NA
#> 249         picalm        V1  -44.2574648   NA
#> 250           pin1        V1    0.6678774   NA
#>  [ reached getOption("max.print") -- omitted 510 rows ]
#> 
#>