Getter for the completed list of nested gene effects

Returns a tibble containing the list of nested genes effects, i.e. a table of nested cluster effects.

nestedGeneEffects(obj)

# S4 method for HMAnalysedPerturbationData
nestedGeneEffects(obj)

Arguments

obj
the object for which you want to extract the underlying effects

Value

returns a tibble.

Examples

 data(rnaiscreen)
 ft <- hm(rnaiscreen)

#> Fitting hierarchical model

 nestedGeneEffects(ft)

#> # A tibble: 1,434 x 5
#>    GeneSymbol Condition  Effect  Qval Control
#>    <chr>      <chr>       <dbl> <dbl>   <int>
#>  1 aak1       HCV       -0.186  1           0
#>  2 abl1       HCV        0.382  1           0
#>  3 abl2       HCV       -0.113  1           0
#>  4 acvr1      HCV       -0.211  1           0
#>  5 acvr1b     HCV        0.178  1           0
#>  6 acvr1c     HCV        0.212  1           0
#>  7 acvr2      HCV        0.0505 1           0
#>  8 acvr2b     HCV        0.246  1           0
#>  9 acvrl1     HCV        0.0136 1           0
#> 10 adam9      HCV       -0.692  0.359       0
#> # … with 1,424 more rows