Create model data for an hierarchical model

Create model data for an hierarchical model

setModelData(obj, drop = TRUE, weights = 1)

# S4 method for PerturbationData
setModelData(obj, drop = TRUE, weights = 1)

Arguments

obj
an data set

drop
boolean if genes that are not found in every Condition should be dropped

weights
a numeric vector used as weights for the single perturbations

Value

returns an PerturbationData object