Getter for graph used for network diffusion

Returns the graph that has been used for analysis using network diffusion.

graph(obj)

# S4 method for NetworkAnalysedPerturbationData
graph(obj)

Arguments

obj
the object for which you want to extract the underlying graph

Value

returns a igraph object

Examples

 data(rnaiscreen)
 ft <- hm(rnaiscreen)

#> Fitting hierarchical model

 gr <- readRDS(system.file(
   "extdata", "graph_small.rds", package = "perturbatr"))
 diffu <- diffuse(ft, gr, r=1)

#> Only taking largest connected component to ensure ergodicity.

#> Diffusion using Markov random walks.

#> Correcting for hub degrees.

#> normalizing column vectors!

#> normalizing column vectors!

 graph(diffu)

#> IGRAPH 6fc6d4f UNW- 549 5202 -- 
#> + attr: name (v/c), weight (e/n)
#> + edges from 6fc6d4f (vertex names):
#>  [1] abi1  --abl1   abi1  --abl2   acvr1 --acvr1b acvr1 --acvr1c acvr1 --acvr2b
#>  [6] acvr1b--acvr2b acvr1c--acvr2b acvr1b--acvrl1 acvr1c--acvrl1 acvr2b--acvrl1
#> [11] acvrl1--adrbk1 adra1a--adrbk1 adra1b--adrbk1 adrb2 --adrbk1 adra1a--adrbk2
#> [16] adra1b--adrbk2 adrb2 --adrbk2 adrbk1--agtr2  akt1  --akt2   akt1  --akt3  
#> [21] akt2  --akt3   acvr1 --amhr2  abl1  --atm    atm   --atr    aurkb --aurkc 
#> [26] avpr1b--bdkrb2 abl1  --blk    blk   --blnk   acvr2b--bmpr1a acvr2b--bmpr1b
#> [31] amhr2 --bmpr1b bmpr1a--bmpr1b acvr1 --bmpr2  acvr1b--bmpr2  acvr2b--bmpr2 
#> [36] acvrl1--bmpr2  bmpr1a--bmpr2  bmpr1b--bmpr2  blnk  --bmx    blk   --btk   
#> + ... omitted several edges